plugin nelder mead test is split up to two tests
