+++ test/plugin_nelder_mead/PluginNelderMeadTests.cpp
@@ -1,0 +1,5 @@
+#include "PluginNelderMeadTests.h"
+
+PluginNelderMeadTests::PluginNelderMeadTests() {
+    pluginsModelsDir = rrTestModelsDir_ / "PLUGINS";
+}
